dzcb.anytone: set Through Mode = On for simplex
On the 578 and 878, DMR MODE = "Simplex" (0)
channels also have "Simplex=On" and
"Through Mode=On" in the exported file.

Neither targeted CPS version exposes
this setting in the UI. But the 578 1.11 CPS
will show DMR MODE "Repeater" unless "Simplex=On"

Se we set it ON for simplex channels, and it
will have zero effect because it only makes a
difference on split channels.