dzcb.repeaterbook: limit default states
Limit default states to Washington and Oregon, since
this is the region where I live and the RPB
API delay is 30 seconds. So avoid a long wait
for 5 * 30 seconds if someone forgets to pass
the state... maybe there shouldn't even be a
default for this value?

Add a TODO about figuring out a smart default.

For the time being, users will probably pass the
states of interest on the command line (generate.sh)