dzcb.model: return to frozen instances
All codeplug objects return to frozen=True and
increase immutability by converting sequence types
to tuple.

This works by excluding `name` and `static_talkgroups`
from the hash function. Two codeplug objects are
equal if all of their corresponding fields are
equal, except the name, because the user might
change that with a regex, etc.

However, since the field is not directly mutable,
we must create a _new_ object with the new name.
And the new object should be __eq__ to the old
object (and compare in sets and dicts as identical).
This fact is used to "update" the container objects,
like grouplist, scanlist, and zone with the newly
renamed channel and contact objects.