dzcb.output.dmrconfig: evolve_from_factory
evolve_from_factory takes on the responsibility of
passing template directives and variables down
to the subtables that get created during initialization.

if the template specifies codeplug filtering, then
apply the filtering to a new table and replace the
existing table in the Dmrconfig_Codeplug. Since
filtering is an expensive operation, avoid filtering
for each subtable