dzcb.dmrconfig: filter contacts in CodeplugIndex
well i can't believe this was the cleanest way
i could find to fix #65 without rewriting a bunch
of stuff. :astonished:

but i think it does an okay job, despite being
obviously hacky.

the CodeplugIndex adds a _contacts_filtered attribute
with init=False that creates a sequence of
contacts from the codeplug unique by name
and preferring the first contact seen. The
contact name -> index mapping is made from this
sequence.

ContactsTable.__iter__ uses the _contacts_filtered
sequence as well when constructing the table.

It's using a private attribute specifically because
it's a deep implementation detail and not a good
design, but is the best place currently to fix it.