dzcb.output.dmrconfig: CodeplugIndexLookup channels_filtered
If the number of channels in the codeplug exceeds
that supported by the radio, reorder the channel
in zone channel order: iterate through the zones
and find all unique channels as they are encountered.

Reordering the channel list prioritizes channels
in preferred zones, but totally ignores any
post-expansion channel-specific Ordering. When true
channel ordering is needed, care should be taken to
stay below the channel limit of the target radio.

fixes #69

update md380-inf.conf expected output to accomodate
the change in the regression test.