dzcb.farnsworth: lookup ContactName when generating
Farnsworth format only allows one contact per ID,
so contacts are collapsed by ID and Slot when
generating the JSON. This process is lossy, resulting
in some channels with talkgroups that have names
which don't appear in the final farnsworth output.

To avoid this issue, we pass a mapping of DMR ID
to Talkgroup when generating channels. Instead
of taking the channel's Talkgroup.name as-is,
look up the ID in the set of known contacts and
use the deduplicated name.